labellab_pipeline-flask/requirements.txt
Flask==1.0.3
python-dotenv-0.12.0
Flask-MongoAlchemy=0.7.2
